+++ src/components/end-user/dashboard/UserTierLevel/components/GradientCircle.tsx
@@ -1,0 +1,11 @@
+// Components
+import LinearGradient from './LinearGradient'
+
+const GradientCircle = () => (
+  <svg fill="none" height="48" style={{ position: 'absolute' }} viewBox="0 0 48 48" width="48">
+    <circle cx="24" cy="24" r="21.5" stroke="url(#paint0_linear_13425_1016)" strokeWidth="5" />
+    <LinearGradient />
+  </svg>
+)
+
+export default GradientCircle
