+++ src/components/end-user/dashboard/UserTierLevel/types.tsx
@@ -1,0 +1,9 @@
+export type TierType = {
+  currentLevel: number
+  currentPoint: number
+  nextLevel: number
+}
+
+export type UserRewards = {
+  quantity?: number
+}
